requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
blinker==1.9.0
    # via flask
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
colorama==0.4.6
    # via click
dnspython==2.7.0
    # via pymongo
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==5.0.1
    # via -r requirements.in
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
polyline==2.0.2
    # via -r requirements.in
pyjwt==2.10.1
    # via -r requirements.in
pymongo==4.11.3
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
urllib3==2.3.0
    # via requests
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
